Add service plugin and support in process integration.
