+++ services/namespaces/service.go
@@ -17,21 +17,13 @@
 package namespaces
 
 import (
-	"strings"
-
-	"github.com/boltdb/bolt"
-	eventstypes "github.com/containerd/containerd/api/events"
 	api "github.com/containerd/containerd/api/services/namespaces/v1"
-	"github.com/containerd/containerd/errdefs"
-	"github.com/containerd/containerd/events"
-	"github.com/containerd/containerd/metadata"
-	"github.com/containerd/containerd/namespaces"
 	"github.com/containerd/containerd/plugin"
+	"github.com/containerd/containerd/services"
 	ptypes "github.com/gogo/protobuf/types"
+	"github.com/pkg/errors"
 	"golang.org/x/net/context"
 	"google.golang.org/grpc"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
 )
 
 func init() {
@@ -39,32 +31,31 @@ func init() {
 		Type: plugin.GRPCPlugin,
 		ID:   "namespaces",
 		Requires: []plugin.Type{
-			plugin.MetadataPlugin,
+			plugin.ServicePlugin,
 		},
 		InitFn: func(ic *plugin.InitContext) (interface{}, error) {
-			m, err := ic.Get(plugin.MetadataPlugin)
+			plugins, err := ic.GetByType(plugin.ServicePlugin)
 			if err != nil {
 				return nil, err
 			}
-			return NewService(m.(*metadata.DB), ic.Events), nil
+			p, ok := plugins[services.NamespacesService]
+			if !ok {
+				return nil, errors.New("namespaces service not found")
+			}
+			i, err := p.Instance()
+			if err != nil {
+				return nil, err
+			}
+			return &service{local: i.(api.NamespacesClient)}, nil
 		},
 	})
 }
 
 type service struct {
-	db        *metadata.DB
-	publisher events.Publisher
+	local api.NamespacesClient
 }
 
 var _ api.NamespacesServer = &service{}
-
-// NewService returns the GRPC namespaces server
-func NewService(db *metadata.DB, publisher events.Publisher) api.NamespacesServer {
-	return &service{
-		db:        db,
-		publisher: publisher,
-	}
-}
 
 func (s *service) Register(server *grpc.Server) error {
 	api.RegisterNamespacesServer(server, s)
@@ -72,158 +63,21 @@ func (s *service) Register(server *grpc.Server) error {
 }
 
 func (s *service) Get(ctx context.Context, req *api.GetNamespaceRequest) (*api.GetNamespaceResponse, error) {
-	var resp api.GetNamespaceResponse
-
-	return &resp, s.withStoreView(ctx, func(ctx context.Context, store namespaces.Store) error {
-		labels, err := store.Labels(ctx, req.Name)
-		if err != nil {
-			return errdefs.ToGRPC(err)
-		}
-
-		resp.Namespace = api.Namespace{
-			Name:   req.Name,
-			Labels: labels,
-		}
-
-		return nil
-	})
+	return s.local.Get(ctx, req)
 }
 
 func (s *service) List(ctx context.Context, req *api.ListNamespacesRequest) (*api.ListNamespacesResponse, error) {
-	var resp api.ListNamespacesResponse
-
-	return &resp, s.withStoreView(ctx, func(ctx context.Context, store namespaces.Store) error {
-		namespaces, err := store.List(ctx)
-		if err != nil {
-			return err
-		}
-
-		for _, namespace := range namespaces {
-			labels, err := store.Labels(ctx, namespace)
-			if err != nil {
-				// In general, this should be unlikely, since we are holding a
-				// transaction to service this request.
-				return errdefs.ToGRPC(err)
-			}
-
-			resp.Namespaces = append(resp.Namespaces, api.Namespace{
-				Name:   namespace,
-				Labels: labels,
-			})
-		}
-
-		return nil
-	})
+	return s.local.List(ctx, req)
 }
 
 func (s *service) Create(ctx context.Context, req *api.CreateNamespaceRequest) (*api.CreateNamespaceResponse, error) {
-	var resp api.CreateNamespaceResponse
-
-	if err := s.withStoreUpdate(ctx, func(ctx context.Context, store namespaces.Store) error {
-		if err := store.Create(ctx, req.Namespace.Name, req.Namespace.Labels); err != nil {
-			return errdefs.ToGRPC(err)
-		}
-
-		for k, v := range req.Namespace.Labels {
-			if err := store.SetLabel(ctx, req.Namespace.Name, k, v); err != nil {
-				return err
-			}
-		}
-
-		resp.Namespace = req.Namespace
-		return nil
-	}); err != nil {
-		return &resp, err
-	}
-
-	if err := s.publisher.Publish(ctx, "/namespaces/create", &eventstypes.NamespaceCreate{
-		Name:   req.Namespace.Name,
-		Labels: req.Namespace.Labels,
-	}); err != nil {
-		return &resp, err
-	}
-
-	return &resp, nil
-
+	return s.local.Create(ctx, req)
 }
 
 func (s *service) Update(ctx context.Context, req *api.UpdateNamespaceRequest) (*api.UpdateNamespaceResponse, error) {
-	var resp api.UpdateNamespaceResponse
-	if err := s.withStoreUpdate(ctx, func(ctx context.Context, store namespaces.Store) error {
-		if req.UpdateMask != nil && len(req.UpdateMask.Paths) > 0 {
-			for _, path := range req.UpdateMask.Paths {
-				switch {
-				case strings.HasPrefix(path, "labels."):
-					key := strings.TrimPrefix(path, "labels.")
-					if err := store.SetLabel(ctx, req.Namespace.Name, key, req.Namespace.Labels[key]); err != nil {
-						return err
-					}
-				default:
-					return status.Errorf(codes.InvalidArgument, "cannot update %q field", path)
-				}
-			}
-		} else {
-			// clear out the existing labels and then set them to the incoming request.
-			// get current set of labels
-			labels, err := store.Labels(ctx, req.Namespace.Name)
-			if err != nil {
-				return errdefs.ToGRPC(err)
-			}
-
-			for k := range labels {
-				if err := store.SetLabel(ctx, req.Namespace.Name, k, ""); err != nil {
-					return err
-				}
-			}
-
-			for k, v := range req.Namespace.Labels {
-				if err := store.SetLabel(ctx, req.Namespace.Name, k, v); err != nil {
-					return err
-				}
-
-			}
-		}
-
-		return nil
-	}); err != nil {
-		return &resp, err
-	}
-
-	if err := s.publisher.Publish(ctx, "/namespaces/update", &eventstypes.NamespaceUpdate{
-		Name:   req.Namespace.Name,
-		Labels: req.Namespace.Labels,
-	}); err != nil {
-		return &resp, err
-	}
-
-	return &resp, nil
+	return s.local.Update(ctx, req)
 }
 
 func (s *service) Delete(ctx context.Context, req *api.DeleteNamespaceRequest) (*ptypes.Empty, error) {
-	if err := s.withStoreUpdate(ctx, func(ctx context.Context, store namespaces.Store) error {
-		return errdefs.ToGRPC(store.Delete(ctx, req.Name))
-	}); err != nil {
-		return &ptypes.Empty{}, err
-	}
-	// set the namespace in the context before publishing the event
-	ctx = namespaces.WithNamespace(ctx, req.Name)
-	if err := s.publisher.Publish(ctx, "/namespaces/delete", &eventstypes.NamespaceDelete{
-		Name: req.Name,
-	}); err != nil {
-		return &ptypes.Empty{}, err
-	}
-
-	return &ptypes.Empty{}, nil
-}
-
-func (s *service) withStore(ctx context.Context, fn func(ctx context.Context, store namespaces.Store) error) func(tx *bolt.Tx) error {
-	return func(tx *bolt.Tx) error { return fn(ctx, metadata.NewNamespaceStore(tx)) }
-}
-
-func (s *service) withStoreView(ctx context.Context, fn func(ctx context.Context, store namespaces.Store) error) error {
-	return s.db.View(s.withStore(ctx, fn))
-}
-
-func (s *service) withStoreUpdate(ctx context.Context, fn func(ctx context.Context, store namespaces.Store) error) error {
-	return s.db.Update(s.withStore(ctx, fn))
+	return s.local.Delete(ctx, req)
 }
