+++ services/snapshots/service.go
@@ -19,17 +19,16 @@ package snapshots
 import (
 	gocontext "context"
 
-	eventstypes "github.com/containerd/containerd/api/events"
 	snapshotsapi "github.com/containerd/containerd/api/services/snapshots/v1"
 	"github.com/containerd/containerd/api/types"
 	"github.com/containerd/containerd/errdefs"
-	"github.com/containerd/containerd/events"
 	"github.com/containerd/containerd/log"
-	"github.com/containerd/containerd/metadata"
 	"github.com/containerd/containerd/mount"
 	"github.com/containerd/containerd/plugin"
+	"github.com/containerd/containerd/services"
 	"github.com/containerd/containerd/snapshots"
 	ptypes "github.com/gogo/protobuf/types"
+	"github.com/pkg/errors"
 	"golang.org/x/net/context"
 	"google.golang.org/grpc"
 )
@@ -39,7 +38,7 @@ func init() {
 		Type: plugin.GRPCPlugin,
 		ID:   "snapshots",
 		Requires: []plugin.Type{
-			plugin.MetadataPlugin,
+			plugin.ServicePlugin,
 		},
 		InitFn: newService,
 	})
@@ -48,20 +47,24 @@ func init() {
 var empty = &ptypes.Empty{}
 
 type service struct {
-	db        *metadata.DB
-	publisher events.Publisher
+	ss map[string]snapshots.Snapshotter
 }
 
 func newService(ic *plugin.InitContext) (interface{}, error) {
-	md, err := ic.Get(plugin.MetadataPlugin)
+	plugins, err := ic.GetByType(plugin.ServicePlugin)
 	if err != nil {
 		return nil, err
 	}
-
-	return &service{
-		db:        md.(*metadata.DB),
-		publisher: ic.Events,
-	}, nil
+	p, ok := plugins[services.SnapshotsService]
+	if !ok {
+		return nil, errors.New("snapshots service not found")
+	}
+	i, err := p.Instance()
+	if err != nil {
+		return nil, err
+	}
+	ss := i.(map[string]snapshots.Snapshotter)
+	return &service{ss: ss}, nil
 }
 
 func (s *service) getSnapshotter(name string) (snapshots.Snapshotter, error) {
@@ -69,7 +72,7 @@ func (s *service) getSnapshotter(name string) (snapshots.Snapshotter, error) {
 		return nil, errdefs.ToGRPCf(errdefs.ErrInvalidArgument, "snapshotter argument missing")
 	}
 
-	sn := s.db.Snapshotter(name)
+	sn := s.ss[name]
 	if sn == nil {
 		return nil, errdefs.ToGRPCf(errdefs.ErrInvalidArgument, "snapshotter not loaded: %s", name)
 	}
@@ -97,12 +100,6 @@ func (s *service) Prepare(ctx context.Context, pr *snapshotsapi.PrepareSnapshotR
 		return nil, errdefs.ToGRPC(err)
 	}
 
-	if err := s.publisher.Publish(ctx, "/snapshot/prepare", &eventstypes.SnapshotPrepare{
-		Key:    pr.Key,
-		Parent: pr.Parent,
-	}); err != nil {
-		return nil, err
-	}
 	return &snapshotsapi.PrepareSnapshotResponse{
 		Mounts: fromMounts(mounts),
 	}, nil
@@ -158,12 +155,6 @@ func (s *service) Commit(ctx context.Context, cr *snapshotsapi.CommitSnapshotReq
 		return nil, errdefs.ToGRPC(err)
 	}
 
-	if err := s.publisher.Publish(ctx, "/snapshot/commit", &eventstypes.SnapshotCommit{
-		Key:  cr.Key,
-		Name: cr.Name,
-	}); err != nil {
-		return nil, err
-	}
 	return empty, nil
 }
 
@@ -178,11 +169,6 @@ func (s *service) Remove(ctx context.Context, rr *snapshotsapi.RemoveSnapshotReq
 		return nil, errdefs.ToGRPC(err)
 	}
 
-	if err := s.publisher.Publish(ctx, "/snapshot/remove", &eventstypes.SnapshotRemove{
-		Key: rr.Key,
-	}); err != nil {
-		return nil, err
-	}
 	return empty, nil
 }
